Implemented CIni::GetBool(), made absolute region coords unsigned.

+++ Ebenezer/EbenezerDlg.h
@@ -119,10 +119,10 @@ public:
 	void NpcInOutForMe(CUser* pSendUser);
 
 	// Get info for NPCs in region
-	void GetRegionNpcIn(C3DMap* pMap, int region_x, int region_z, Packet & pkt, uint16 & t_count);
+	void GetRegionNpcIn(C3DMap* pMap, uint16 region_x, uint16 region_z, Packet & pkt, uint16 & t_count);
 
 	// Get list of NPC IDs in region
-	void GetRegionNpcList(C3DMap* pMap, int region_x, int region_z, Packet & pkt, uint16 & t_count);
+	void GetRegionNpcList(C3DMap* pMap, uint16 region_x, uint16 region_z, Packet & pkt, uint16 & t_count);
 
 	// Get list of NPCs for regions around user (WIZ_NPC_REGION)
 	void RegionNpcInfoForMe(CUser* pSendUser);	
@@ -131,10 +131,10 @@ public:
 	void UserInOutForMe(CUser* pSendUser);
 
 	// Get info for users in region
-	void GetRegionUserIn(C3DMap* pMap, int region_x, int region_z, Packet & pkt, uint16 & t_count);
+	void GetRegionUserIn(C3DMap* pMap, uint16 region_x, uint16 region_z, Packet & pkt, uint16 & t_count);
 
 	// Get list of user IDs in region
-	void GetRegionUserList(C3DMap* pMap, int region_x, int region_z, Packet & pkt, uint16 & t_count);
+	void GetRegionUserList(C3DMap* pMap, uint16 region_x, uint16 region_z, Packet & pkt, uint16 & t_count);
 
 	// Get list of users for regions around user (WIZ_REGIONCHANGE)
 	void RegionUserInOutForMe(CUser* pSendUser);
@@ -143,7 +143,7 @@ public:
 	void MerchantUserInOutForMe(CUser* pSendUser);
 
 	// Get list of merchants in region
-	void GetRegionMerchantUserIn(C3DMap* pMap, int region_x, int region_z, Packet & pkt, uint16 & t_count);
+	void GetRegionMerchantUserIn(C3DMap* pMap, uint16 region_x, uint16 region_z, Packet & pkt, uint16 & t_count);
 
 	__forceinline bool isPermanentMessageSet() { return m_bPermanentChatMode == TRUE; }
 	void GetPermanentMessage(Packet & result);
